chore: Add start and stop scripts for managing services
This commit adds two new scripts, start.sh and stop.sh, to the repository. The start.sh script is used to start the services using docker-compose, while the stop.sh script is used to stop the services. These scripts provide a convenient way to manage the services without having to manually run the docker-compose commands.